Refactor style editor with modular architecture and plugin system
